requirements.txt
# multi_personas_agent/requirements.txt
google-adk
python-dotenv